requirements.txt
pytorch=2.4.0
torchaudio=2.2.0
torchvision=0.19.0
transformers=4.44.2
boto3=1.35.10
streamlit=1.38.0
